refactor: remove redundant selector of auth slice
